transaction 추가 및 update 수정

+++ src/main/java/com/tenco/hobby/service/BoardService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.tenco.hobby.dto.CommentDto;
 import com.tenco.hobby.dto.UpdateFormDto;
@@ -19,6 +20,7 @@ import com.tenco.hobby.repository.model.BoardHobbies;
 import com.tenco.hobby.repository.model.Comment;
 import com.tenco.hobby.repository.model.Hobby;
 import com.tenco.hobby.repository.model.Report;
+import com.tenco.hobby.repository.model.User;
 
 @Service
 public class BoardService {
@@ -35,6 +37,7 @@ public class BoardService {
 	/**
 	 * 전체 글조회
 	 */
+	@Transactional
 	public List<Board> readBoardList() {
 		List<Board> list = boardRepository.findAll();
 		return list;
@@ -42,9 +45,9 @@ public class BoardService {
 
 	/**
 	 * 글선택조회
-	 * 
 	 * @param id
 	 */
+	@Transactional
 	public Board readBoard(Long id) {
 
 		Board boardEntity = boardRepository.findById(id);
@@ -52,9 +55,9 @@ public class BoardService {
 	}
 
 	/**
-	 * 
 	 * @return 취미 전체조회
 	 */
+	@Transactional
 	public List<BoardHobbies> readHobbyCategory() {
 		List<BoardHobbies> list = hobbyRepository.findAllHobbies();
 		return list;
@@ -64,6 +67,7 @@ public class BoardService {
 	 * @param id
 	 * @return 취미카테고리 글 조회
 	 */
+	@Transactional
 	public List<Board> readHobbyList(Long id) {
 		List<Board> list = boardRepository.findByHobbyId(id);
 		return list;
@@ -71,10 +75,10 @@ public class BoardService {
 
 	/**
 	 * 댓글 조회
-	 * 
 	 * @param boardId
 	 * @return comment
 	 */
+	@Transactional
 	public List<Comment> readComment(Long boardId) {
 
 		List<Comment> commentList = commentRepository.findByBoardId(boardId);
@@ -87,6 +91,7 @@ public class BoardService {
 	 * @param writeFormDto
 	 * @param principalId
 	 */
+	@Transactional
 	public void createPost(WriteFormDto writeFormDto, Long principalId) {
 
 		Board board = new Board();
@@ -107,6 +112,7 @@ public class BoardService {
 	 * @param principalId
 	 * @param boardId
 	 */
+	@Transactional
 	public void createComment(CommentDto commentDto, Long principalId, Long boardId) {
 
 		Comment comment = new Comment();
@@ -126,15 +132,23 @@ public class BoardService {
 	 * @param principalId
 	 * @param id
 	 */
+	@Transactional
 	public void updatePost(UpdateFormDto updateFormDto, Long principalId, Long id) {
 
-		Board board = new Board();
-		board.setTitle(updateFormDto.getTitle());
-		board.setContent(updateFormDto.getContent());
-		board.setUserId(principalId);
-		board.setHobbyId(updateFormDto.getHobbyId());
-		board.setId(id);
-		int resultRowCount = boardRepository.updateById(board);
+		Board boardEntity = boardRepository.findById(id);
+		if (boardEntity == null) {
+			throw new CustomRestfullException("해당 글이 존재하지 않습니다", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+
+		if (principalId != boardEntity.getUserId()) {
+			throw new CustomRestfullException("수정 권한이 없습니다", HttpStatus.BAD_REQUEST);
+		}
+		boardEntity.setTitle(updateFormDto.getTitle());
+		boardEntity.setContent(updateFormDto.getContent());
+		boardEntity.setUserId(principalId);
+		boardEntity.setHobbyId(updateFormDto.getHobbyId());
+		boardEntity.setId(id);
+		int resultRowCount = boardRepository.updateById(boardEntity);
 		if (resultRowCount != 1) {
 			throw new CustomRestfullException("글수정 실패하였습니다", HttpStatus.INTERNAL_SERVER_ERROR);
 		}
@@ -147,12 +161,21 @@ public class BoardService {
 	 * @param commentDto
 	 * @param id
 	 */
-	public void updateComment(CommentDto commentDto, Long id) {
+	@Transactional
+	public void updateComment(CommentDto commentDto, Long id, Long principalId) {
 
-		Comment comment = new Comment();
-		comment.setContent(commentDto.getContent());
-		comment.setId(id);
-		int resultRowCount = commentRepository.updateById(comment);
+		Comment commentEntity = commentRepository.findById(id);
+		if (commentEntity == null) {
+			throw new CustomRestfullException("해당 댓글이 존재하지 않습니다", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+
+		if (principalId != commentEntity.getUserId()) {
+			throw new CustomRestfullException("수정권한이 없습니다", HttpStatus.BAD_REQUEST);
+		}
+
+		commentEntity.setContent(commentDto.getContent());
+		commentEntity.setId(id);
+		int resultRowCount = commentRepository.updateById(commentEntity);
 		if (resultRowCount != 1) {
 			throw new CustomRestfullException("댓글수정에 실패하였습니다", HttpStatus.INTERNAL_SERVER_ERROR);
 		}
@@ -160,10 +183,19 @@ public class BoardService {
 
 	/**
 	 * 글삭제
-	 * 
 	 * @param id
 	 */
-	public void deletePost(Long id) {
+	@Transactional
+	public void deletePost(Long id, Long principalId) {
+
+		Board boardEntity = boardRepository.findById(id);
+		if (boardEntity == null) {
+			throw new CustomRestfullException("해당 글이 존재하지 않습니다", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+
+		if (principalId != boardEntity.getUserId()) {
+			throw new CustomRestfullException("접근 권한이 없습니다", HttpStatus.BAD_REQUEST);
+		}
 
 		int resultRowCount = boardRepository.deleteById(id);
 		if (resultRowCount != 1) {
@@ -173,10 +205,20 @@ public class BoardService {
 
 	/**
 	 * 댓글 삭제
-	 * 
 	 * @param id
 	 */
-	public void deleteComment(Long id) {
+	@Transactional
+	public void deleteComment(Long id, Long principalId) {
+		
+		Comment commentEntity = commentRepository.findById(id);
+		if (commentEntity == null) {
+			throw new CustomRestfullException("해당 댓글이 존재하지 않습니다", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		if (principalId != commentEntity.getUserId()) {
+			throw new CustomRestfullException("접근 권한이 없습니다", HttpStatus.BAD_REQUEST);
+		}
+		
 		int resultRowCount = commentRepository.deleteById(id);
 		if (resultRowCount != 1) {
 			throw new CustomRestfullException("댓글삭제 실패하였습니다", HttpStatus.INTERNAL_SERVER_ERROR);
@@ -185,10 +227,10 @@ public class BoardService {
 
 	/**
 	 * 게시글 신고
-	 * 
 	 * @param id
 	 * @param principalId
 	 */
+	@Transactional
 	public void createReportPost(Long id, Long principalId) {
 
 		Report report = new Report();
@@ -201,6 +243,7 @@ public class BoardService {
 		}
 	}
 
+	@Transactional
 	public void createReportCmt(Long id, Long principalId) {
 
 		Report report = new Report();
